[chore #163911078] politico version 2 project set up
