+++ requirements.txt
@@ -10,7 +10,9 @@ Jinja2==2.10
 MarkupSafe==1.1.0
 more-itertools==5.0.0
 pluggy==0.8.1
+psycopg2==2.7.7
 py==1.7.0
+pycodestyle==2.5.0
 pytest==4.2.0
 pytest-cov==2.6.1
 pytz==2018.9
